+++ frontend/src/pages/login/Login.tsx
@@ -2,13 +2,26 @@ import React from 'react';
 import { useLocation } from 'react-router-dom';
 import Photo from '../login/login-logo.png';
 import styles from '../login/login.module.css';
-import { GitHubURL } from '../../client/models/GithubURL';
 
 type LocationState = {
   from: {
     path: string;
   };
 };
+
+function GitHubURL(from: string) {
+  const rootUrl = 'https://github.com/login/oauth/authorize';
+  const options = {
+    client_id: '' as string,
+    redirect_url: 'http://localhost:8080/login' as string,
+    scope: 'user:email',
+    state: from,
+  };
+
+  const qs = new URLSearchParams(options);
+
+  return `${rootUrl}?${qs.toString()}`;
+}
 
 const Login = () => {
   const location = useLocation();
